Feature be/barter dto map struct

See merge request s09-fintech-finance-sub2/S09P22B207!147

+++ Back-End/src/main/java/com/conseller/conseller/store/dto/response/DetailStoreResponse.java
@@ -18,7 +18,7 @@ public class DetailStoreResponse {
 
     private String storeUserProfileUrl;
 
-    private Integer storeUserDeposit;
+    private Long storeUserDeposit;
 
     private Long storeIdx;
